requirements.txt
ops>2,<4
pyyaml
pydantic